backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Flask-Session==0.6.0
Flask-CORS==4.0.0
cachetools==5.3.2
numpy<2.0.0
chromadb>=0.4.0,<0.5.0
sentence-transformers
google-generativeai>=0.3.0
langchain-google-genai
langgraph
